interface/requirements.txt
pypdf
pillow
streamlit
firebase_admin